pyproject.toml
[tool.poetry]
name = "airq"
version = "0.1.0"
description = ""
authors = ["RavshanjonA <akhmadjonovr.98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
ruff = "^0.8.1"
django-filter = "^24.3"
google-api-python-client = "^2.154.0"
facebook-sdk = "^3.1.0"
drf-standardized-errors = "^0.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
